[perf, refactor] 로그인 페이지 웹 접근성 개선 및 아키텍쳐 리팩토링 - 로그인 버튼 색상 대비 개선 - 폼 라벨 연결 - 스크린 리더 지원 - 에러 메시지 접근성 - features 구조 적용 - 관심사 분리
